app/requirements.txt
attrs==19.1.0
falcon==2.0.0
gunicorn==19.9.0
jsonschema==3.0.1
marshmallow==2.19.5
psycopg2==2.8.3
webargs==6.1.0
pytest==5.4.3
PyJWT==1.7.1